Analysis Summary can be now written both as a text and HTML at the same time
